Remove types.js, move ts_types.ts -> types.ts
