+++ src/log/index.ts
@@ -5,7 +5,7 @@ import {
   ProjectSource,
   ThunkConfig,
   LogStatus,
-} from '../ts_types'
+} from '../types'
 
 let id = 0
